# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanceFlow API",
        "description": "FinanceFlow API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server development (localhost)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server development (IP)"
        }
    ],
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get a list of all user accounts",
                "operationId": "4bbac367ecf8af0f1c2dbad57ff21dc3",
                "parameters": [
                    {
                        "name": "archived",
                        "in": "query",
                        "description": "Set to true to include archived accounts",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of accounts"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Create a new account",
                "operationId": "ca10abbfc49106114ac7da3d8678501b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "currency_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Main Card"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "card",
                                        "enum": [
                                            "card",
                                            "cash",
                                            "crypto",
                                            "deposit",
                                            "investment"
                                        ]
                                    },
                                    "currency_code": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#3B82F6",
                                        "nullable": true
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "credit-card",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get a specific account by ID",
                "operationId": "577c024ccebfc85ec3189eb045d43f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Soft-delete an account (cascades to transactions)",
                "operationId": "92228661a3432dd364690c18850eb0a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Account deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update a specific account",
                "operationId": "ecbb59c4c8754ace8def38c009e54d06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "card",
                                        "enum": [
                                            "card",
                                            "cash",
                                            "crypto",
                                            "deposit",
                                            "investment"
                                        ]
                                    },
                                    "currency_code": {
                                        "type": "string",
                                        "example": "EUR"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#10B981",
                                        "nullable": true
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "wallet",
                                        "nullable": true
                                    },
                                    "is_archived": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/attachments": {
            "post": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Upload a file and attach it to a transaction",
                "description": "Accepted formats: jpg, jpeg, png, gif, webp, pdf. Maximum file size: 5 MB.",
                "operationId": "e72af5df2d4d5f20cfbb99526ab8ff14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "transaction_id",
                                    "file"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attachment uploaded"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error (wrong format or size)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/attachments/{id}": {
            "delete": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Delete an attachment and remove the file from storage",
                "operationId": "9d090bf09bcab12eef937baddc7176c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Attachment deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}